allow some YInput types to be manually disposed
This allows us to initialise a collection of YInputs, consume their underlying gen.YInput's, then force the Dart GC to keep the Dart YInput alive until we're done. If we have a reference to a Dart YInput, as required to call dispose(), it implies that the object should not be GC'd.